File reorganization.
