+++ src/components/Image/Image.css
@@ -1,0 +1,5 @@
+.image {
+    height: 100%;
+    width:100%;
+
+}
