+++ src/components/NavBar/Navbar.css
@@ -1,51 +1,37 @@
 .navbar {
     display: flex;
-    justify-self: start;
-    justify-content:start;
+    align-items: center;
+    justify-content:space-between;
     position: fixed;
-    background-color: #FFFFFF;
-    height: 10vh;
-    width: 100%;
 }
 
-.menu-icon {
-    width: 80%; 
+.menu-icon { 
     margin-left: 2rem;
     margin-right: 2rem;
     font-size: 2rem;
     text-decoration: none;
     color: black;
+    transform: rotate(-90deg);
 }
 
-.social-bar {
-    display: flex;
-    gap: 2rem;
-    margin-right: 2rem;
-    align-items: baseline; 
-  }
-
-.social-icon {
-    color: #FFFFFF;
-  }
-
-
 .nav-menu {
-    width: 100%;
-    height: 100%;
     display: flex;
     justify-content: center;
     position:absolute;
-    z-index: 1;
+    z-index: 200;
     top: 0;
-    left: -100%; 
-    transition: 850ms;
-    background-color: #FFFFFF;
-
+    background-color: #D3CBBD;
+    transform: translateX(-100%);
+    transition: transform 850ms ease-out;
+    width: 100%;
+    height: 100%;
 }
 
 .nav-menu.active {
-    left: 0;
-    transition: 850ms;
+    width: 100%;
+    height: 100%;
+    right: 0;
+    transform: translateX(0);
     
 }
 
@@ -56,8 +42,6 @@
     padding: 20px 0px 20px 16px;
     list-style: none;
     height: 10vh;
-    color: black;
-    
 }
 
 .nav-text a {
@@ -66,6 +50,7 @@
     padding: 0 2rem;
     transition: 0.3s;
     animation: ease-in-out;
+    color: black;
 }
 
 .nav-text a:hover {
@@ -82,8 +67,7 @@
 }
 
 .navbar-toggle {
-    width: 100%;
-    height: 80px;
+    width: 100vw;
     display: flex;
     justify-content: flex-start;
     align-items: center;
