--- src/components/Socialbar.css
@@ -1,7 +1,0 @@
-.social-icons {
-    display: flex;
-    color: #000000;
-    flex-direction: column;
-    list-style: none;
-    text-decoration: none;
-  }
