--- src/components/Socialbar.js
@@ -1,27 +1,0 @@
-import './Socialbar.css';
-import { TiSocialLinkedin } from 'react-icons/ti';
-import { TiSocialGithub } from 'react-icons/ti';
-import { TiSocialTwitter} from 'react-icons/ti';
-import { CgMail } from 'react-icons/cg';
-import React, {useState} from 'react';
-import { Link } from 'react-router-dom';
-
-<div>
-  <ul className="social-icons">
-          <li >
-              <CgMail />
-          </li>
-          <li>
-              <TiSocialGithub />
-          </li>
-          <li>
-              <TiSocialLinkedin />
-          </li>
-          <li>
-              <TiSocialLinkedin />
-          </li>
-          <li>
-              <TiSocialTwitter />
-          </li>
-      </ul>
-  </div>
