+++ src/components/TextSection/TextSection.css
@@ -1,0 +1,3 @@
+.content {
+    margin: 2rem 2rem;
+}
