+++ src/components/TextSection/TextSection.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import './TextSection.css'
+
+var TextSection= (props) => {
+    return(
+        <p className="content">{props.content}</p> 
+    )
+}
+
+export default TextSection;
